# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuizApp API Documentation",
        "contact": {
            "email": "haykalcs@gmail.com"
        },
        "version": "1"
    },
    "paths": {
...
        "/api/feeds": {
            "get": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Get List of Feeds",
                "description": "Get List of Feeds",
                "operationId": "getFeedsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "message": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "user": {
                                                        "type": "object"
                                                    },
                                                    "replies": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "feed_id": {
                                                                    "type": "integer"
                                                                },
                                                                "user_id": {
                                                                    "type": "integer"
                                                                },
                                                                "message": {
                                                                    "type": "string"
                                                                },
                                                                "image": {
                                                                    "type": "string"
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "format": "date-time"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "format": "date-time"
                                                                },
                                                                "user": {
                                                                    "type": "object"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "user_id": 1,
                                                    "message": "Hi there!",
                                                    "image": "2079572408.jpg",
                                                    "created_at": "2021-10-10T12:05:19.000000Z",
                                                    "updated_at": "2021-10-10T12:05:19.000000Z",
                                                    "user": {
                                                        "id": 1,
                                                        "name": "John Doe",
                                                        "email": "johndoe@gmail.com"
                                                    },
                                                    "replies": [
                                                        {
                                                            "id": 1,
                                                            "feed_id": 1,
                                                            "user_id": 1,
                                                            "message": "Hallo there!",
                                                            "image": "2093156899.jpg",
                                                            "created_at": "2021-10-10T12:07:08.000000Z",
                                                            "updated_at": "2021-10-10T12:07:08.000000Z",
                                                            "user": {
                                                                "id": 1,
                                                                "name": "Mike",
                                                                "email": "mike@gmail.com"
                                                            }
                                                        }
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Store Feed in DB",
                "description": "Store Feed in DB",
                "operationId": "storeFeed",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Hi there!"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "nullable": [
                                    "image"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data berhasil dibuat"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "user_id": 2,
                                                "message": "Hi there!",
                                                "image": "1172193584.jpg",
                                                "updated_at": "2021-09-26T13:00:59.000000Z",
                                                "created_at": "2021-09-26T13:00:59.000000Z",
                                                "id": 3
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/feeds/{feedId}/reply": {
            "post": {
                "tags": [
                    "Feeds"
                ],
                "summary": "Store Reply in DB",
                "description": "Store Reply in DB",
                "operationId": "storeReply",
                "parameters": [
                    {
                        "name": "feedId",
                        "in": "path",
                        "description": "Id of Feed",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Hallo there!"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "nullable": [
                                    "image"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data berhasil dibuat"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "user_id": 2,
                                                "feed_id": 1,
                                                "message": "Hallo there!",
                                                "image": "1172193584.jpg",
                                                "updated_at": "2021-09-26T13:00:59.000000Z",
                                                "created_at": "2021-09-26T13:00:59.000000Z",
                                                "id": 4
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}